replace doc w/ a transaction, instead of generating new state
Prosemirror doesn't want us to generate an entire new editor state;
this causes problems like blowing away plugin state.
(https://discuss.prosemirror.net/t/replacing-a-states-doc/634/12)

The more correct approach is to generate a transaction that replaces the
entire doc contents. This commit switches us over to that model.

(Of course, generating minimal transactions that don't replace the
entire doc may be even more preferable, but this is a start.)