Fix handling of local sequence numbers
Previously we had an issue where the following sequence
would cause an error:

- Alice generates a local change
- Bob generates a change depending on Alice's change
- Alice tries to apply Bob's change, but the code thinks it's missing a dependency: the original change generated by Alice herself.

The problem was that local change generation wouldn't
update the sequence number for the local actor in the clock,
until the next change was generated.

This commit changes that behavior, so that
we update the clock for the local actor when we
generate a local change.

So far, have only tested:

- this change fixes our editor UI demo.
- existing minimal test case for Micromerge passes.

Todo: write a minimal test case for this behavior in isolation.