correctly maintain format span metadata on text insertion/deletion
We currently store format spans in Micromerge using integer indexes.

However, previously we were not actually updating these integer indexes when characters were inserted/deleted in the document, so formatting wouldn't stay attached to the text properly.

This commit adds code that updates the format span indexes upon insertion/deletion operations.

I'm not sure yet this is necessarily the best approach... perhaps we should instead be storing format spans with character IDs, and then resolving those to integers at a later time? In the end all the approaches are kind of equivalent, it's just a matter of weighing which style is likely to be most robust and also efficient.

This approach is a nice demonstration of how Micromerge is useful, because we can easily change how text insertion interacts with formatting.