An incomplete attempt at removing +1/-1 arithmetic
This commit is a half-baked attempt with lots of failing tests.
The goal was to remove math that increments/decrements integer
character indexes upon character insertion/deletion, but it
didn't work out as cleanly as I'd hoped.

Let's think about the history of how we got here.
Originally when prototyping on top of Automerge, we were
only maintaining a list of format operations, and operations
on the text itself were managed by Automerge.

Our original algorithm was designed to fully replay all ops
in some order from 0. Ops would store span information using
character IDs; we'd resolve all the char IDs into integer indexes
based on whatever the current string content is, and then
play all the ops.

Then we started moving into a more "out-of-order" application style,
where we maintain some intermediate state. In particular,
we would keep around a "list of flattened spans" and figure out how
to apply any new op to that list of spans, producing a
new list of spans. We tested this and it worked great.

However I think we missed a case. What happens if the text string
gets updated? The intermediate state is represented in terms of
integer indexes, which means it's no longer aligned with the new text.
This is bad! Here are two options to fix the situation:

1) When text is inserted/deleted, update the intermediate state.
   In particular, update the integer indexes for spans after the
   insertion/deletion. This is the approach that was in the code
   before this commit. It works fine for now. However we are worried
   that it may end up leading to bugs/confusion later.

2) Change the representation of the intermediate state.
   Rather than storing integer indexes, store the flattened list of
   spans in terms of character IDs. This seems more robust and
   avoids the need for any integer index updates. This was the idea
   to try in this commit. However as I went, I realized:
   this seems really inefficient? The problem is that all our
   format application logic is defined in terms of integer indexes.
   So every time we apply a format op we do the following dance:
   convert char IDs for all spans to integer indexes, apply the
   formatting in int-space, and then convert back to char-ID-space.
   Even with efficient structures to go in both directions this seems
   like a lot of work to do to apply a single operation.

This commit got partway through implementing that Option 2.
I don't know why the tests aren't passing yet, but it's probably
doable to make it work. It just seems really inefficient.

Maybe a better solution would be to instead rework our formatting
application logic to work directly in char-ID-space, so there's no
need to go back and forth between char IDs and numbers?
Still need to think about how feasible this would be.

Or we can stick with the current +1/-1 integer index updating
and see how that goes.

I think an important thing to note is that ops are the source of truth,
and all ops are defined in terms of char IDs;
these integer indexes are only present as a sort of caching structure
to help us maintain state as we apply ops.
So that may be a small vote in favor of sticking with an approach
that allows for some integer indexes; at least we'll never end up
in a situation where we've persisted corrupt integer indexes?