Fix root doc type errors
Peritext can't annotate the type of the root document because it is
inherently mutable. Unfortunately, this means that callers accessing the
document via `doc.root` also have no static guarantees about the shape
of the document.

As an Elmer's glue fix, we expose a method `getRoot` to allow callers to
pass in a type parameter dictating the shape of the root document. As a
halfhearted attempt to impose some kind of safety, we return a document
with all the top-level fields optional so that the caller still has
to perform a modicum of runtime checking.