Replace Automerge with Micromerge (#5)
This change replaces Automerge with Micromerge, a minimal CRDT implementation with types.

* Split logic into bridge.ts

Create a new file bridge.ts for handling all logic between ProseMirror
<> CRDT.

- Remove Automerge and dependencies
- Remove global variables

* Implement Publisher interface

* Enqueue changes to send to publisher

* Fix handling of local sequence numbers

Previously we had an issue where the following sequence
would cause an error:

- Alice generates a local change
- Bob generates a change depending on Alice's change
- Alice tries to apply Bob's change, but the code thinks it's missing a dependency: the original change generated by Alice herself.

The problem was that local change generation wouldn't
update the sequence number for the local actor in the clock,
until the next change was generated.

This commit changes that behavior, so that
we update the clock for the local actor when we
generate a local change.

So far, have only tested:

- this change fixes our editor UI demo.
- existing minimal test case for Micromerge passes.

Todo: write a minimal test case for this behavior in isolation.

* Extract Micromerge tests

- Moves Micromerge tests out to a file in our test suite
- Adds a test for the dependencies bug we saw yesterday;
  confirmed the test fails before our fix and passes after.
- Switches the Micromerge source to be a Typescript file
  with a "nocheck" comment at the top to disable errors.
  This was the easiest way to get module export stuff
  working with the tests. Very soon, we'll properly
  add types.

* initial sketch: applying formatting ops in micromerge

This commit removes the formatting logic that was
previously in Micromerge, and replaces it with our
implementation of format spans. Some key concepts:

- For each Text object (ie, list of characters), we store
  a flattened list of format spans associated with that text.
- Whenever we receive an addMark or removeMark operation,
  we apply it directly to the flattened list of format spans.
  Because of our previous work on handling out-of-order ops,
  it's always safe to apply the latest op to the flattened spans;
  by comparing opIDs and doing LWW on format attributes, we ensure
  that we'll converge to the same result across peers even though
  ops get applied in a different order.
- We expose an external API from Micromerge which returns
  "format spans with text" -- a list of formatted spans, with the
  text content included inline. The intent here is convenience, so
  the library user doesn't need to manually slice into a string to
  reconstruct the final formatted text.

There are some basic test cases to ensure things are roughly working.

remaining todos:

- add formatting operations back to our UI demo so we can test full stack
- add types throughout Micromerge, fix all type errors
- add more unit tests

* Add types for Operation and Metadata

Start adding types to micromerge.

- Operation: primarily InputOperations used in tests; still need to
define internal Operation types with `opId` etc.
- Metadata: this isn't typechecking currently, but it shows the types of
`this.metadata` used for tracking "managed" objects and arrays.

* Fix metadata typechecking

I forgot that a string index signature will always override any explicit
string keys, so

```typescript
type Foo = { [x: string]: number } & { children: boolean }
```

is equivalent to

```typescript
type Foo = { [x: string]: number }
```

But making the `children` key a symbol (which we should do, anyway!)
fixes this problem.

* Fix root object typechecking

* findListElement and getListElementId

* getObjectIdForPath

* List method refactoring

* recomputeFormatting

* Add types for internal operations

- Introduce types for internal operations, distinct from input
operations
- Refactor methods to exhaustively refine operations by action
- Refactor surrounding files to use new formatting operations

* Update format tests

- Use `action` instead of `type`
- Pass `opId` as a string instead of a number

* Bring back getTextWithFormatting

* Fix metadata temporary

* Integrate Micromerge with CRDT and bridge

* fix deps bug + formatted text output

* add cursors, fix deletion bug

* keep span indexes updated upon insertion/deletion

* Fix root doc type errors

Peritext can't annotate the type of the root document because it is
inherently mutable. Unfortunately, this means that callers accessing the
document via `doc.root` also have no static guarantees about the shape
of the document.

As an Elmer's glue fix, we expose a method `getRoot` to allow callers to
pass in a type parameter dictating the shape of the root document. As a
halfhearted attempt to impose some kind of safety, we return a document
with all the top-level fields optional so that the caller still has
to perform a modicum of runtime checking.

Co-authored-by: Geoffrey Litt <gklitt@gmail.com>