Initial bad implementation of comments
A partial stab at implementing comments as new kind of mark.
Introduces "hasIdentity" as a boolean config value on marks,
indicating that the marks have identity and carry data.

Useful for seeing all the places that need to change,
but has tons of problems. Lots of messiness around the
two different kinds of marks interacting, lots of type errors.