Return MarkValues as-is in getTextWithFormatting
Previously we'd return a simplified representation in getTextWithFormatting:
a boolean value for each mark type.

Now that we have different kinds of mark values, it seems simplest to
just return the mark values directly, but omitting OpIds (because
op ids are an internal consideration, not for external consumers).
So for bold/italic we return `{active: true}` in place of `true`.

We also have to update our tests to accommodate the new structure.