Define attrs type for operations
Some marks have attrs, such as `comment` which stores the comment `id`.
Until now, attrs was typed as `any` on operations. This commit primarily
adds proper typing for attrs based on a type-level map from `MarkType`
to mark values.

We also monomorphized the Micromerge implementation to refer to
`MarkType` and `MarkValue` directly, in order to simplify the code. This
requires removing the type parameter from several other files
transitively referencing the Micromerge class.