roundtrip local editor changes thru Micromerge
Previously we'd directly apply Prosemirror transactions in the local editor.
This increased the risk of divergence between peers, since
Micromerge's interpretation of the change was irrelevant locally.

Now we do something safer: roundtrip the Prosemirror step through Micromerge locally.
We generate a MM change, create a patch from applying the change,
and then we turn that patch back into a PM tranaction that we then
apply locally.

Now both local/remote editors are based on the Micromerge state,
which substantially reduces the risk of diverging editors.