include marks at the given span position in incremental insert step
When we tell Prosemirror to insert a character, we need to explicitly instruct it
what marks should be applied to that character.

To do this, we lookup the span that's active at that position, observe the marks
on that span, and then include them in the insert step that we send to Prosemirror.
This ensures that newly added characters have the correct formatting from the
span that they are inside of.