Fix type error
TypeScript couldn't prove that the variable `patch` was assigned in all
reachable branches. Now, we annotate the type of `unreachable` to return
`never`, and `return unreachable(x)` instead of just calling
`unreachable(x)`. This allows the typechecker to check all branches
exhaustively and prove that `patch` is assigned in all reachable
codepaths.