correctly handle ops that start/end on same character
Previously we incorrectly assumed that an op would
start and end on different characters.
This led to the effect of ops being extended beyond
the end of the op.
Now we correctly handle this case by checking for
whether an op is ending on an element, even if it
also started on that same element.