fix handling of ops on end of a span
When an op applies to only the last character in a span,
we previously didn't handle it properly.
We would retrieve the existing ops on that character
and remove any ops that ended on that character--
this is indeed what we want to do if we're getting
existing ops for a character to the right of some
character with existing ops in the metadata, but
it's not the correct behavior when the ops are actually
on the character we're looking at.

This commit fixes that behavior.