Fix patch emission code in boundary cases
We have code that emits incremental patches describing the
effects of an incoming op, while applying the op.
If there are already format regions present where the op is
being applied, it might have an effect in some parts but
not in others. For example, an op that bolds text between
characters 1 through 10 might result in characters 1-3 and
7-10 being bolded, but not 4-6.

As a result, we need to observe where formatting ops have
already been applied, and only emit patches describing changes
in the appropriate spans.

Previously, our code made a couple incorrect assumptions.
The main one was that a single
character can only serve as the beginning or end of a span.
But this isn't true--a single character can be both the beginning
and the end of a span; the simplest case is just a single bolded
character which is both the start and end of a span.

We've now removed that assumption and added some simple test cases
which confirm that patches are correctly emitted when format ops
overlap with each other.