fix off-by-one error
We had a bug in our code for finding the closest set of mark ops to the left of a position.
If we were looking on the "before" position of a character at index i, we would first
look at the "after" position of that same character before moving leftwards.
This is obviously incorrect; we only want to copy from positions to the left of the
given position.

This happened to manifest as a bug when two links were concurrently created,
one of them was a single character long, and they both ended at the same position.

We now have a test case for that specific case, and this commit fixes the bug.