disable insert-after-tombstone feature
We had a feature where sometimes we insert characters after tombstones
to help deal with non-growing spans which have their boundary chars deleted.
That feature was leading to de-syncs.

If we remove the feature (and skip its corresponding test for now),
fuzzing seems to pass.

I don't yet know exactly why the feature breaks. All of the fuzzer
traces that trigger it have been pretty large.
links-nice.json is the smallest trace I could find which demonstrates
this failure mode; it was collected before removing the feature.