Fix equality checks on comments marks
Previously we had spurious test/fuzzer failures involving comment marks.
The problem was that comments on a span are essentially a Set,
but we currently store them in an array, which makes some
equality checks fail because things are repeated or out of order.

For now, this fixes those problems by making sure to:
- only add a comment to a span if it's not already there
- sort lists of comments on a span by ID when we write to the list,
  to make equality checks pass between equivalent sets of comments

Probably a better refactor would be to switch to storing these in a Set,
but that's a nontrivial refactor so I'm not doing that yet for now.